requirements.txt
fastapi
uvicorn
python-dotenv
langchain_qdrant
langchain_huggingface
langchain_google_genai
langchain_text_splitters
langchain_community
qdrant-client
langchain
langchain_classic
sentence-transformers
pypdf
pillow
pytesseract
pdf2image
sqlalchemy
pydantic
python-multipart
google-genai